requirements_uploader.txt
# requirements_uploader.txt
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
ollama
pysubs2
curses-menu
tqdm
resemble-perth==1.0.1
instagrapi
python-dotenv
